Moving Transform: Refactor -- extract aggregations and periods
